+++ CHANGELOG.md
@@ -5,6 +5,7 @@
     - Fix expired session token renewal when polling results (snowflakedb/snowflake-jdbc#2489)   
     - Fix missing minicore async initialization that was dropped during public API restructuring in v4.0.0
     - Adjust level of logging during Driver initialization
+    - Add sanitization for nonProxyHosts RegEx patterns
 
 - v4.0.1
     - Add /etc/os-release data to Minicore telemetry